Tested
Little fix to the material

Loader of rungholt work in progress. Correctly fill materials and load textures, I need to write the correct shader. Also, I definitely need to cache textures in the texture manager class, because right now I'm loading dulicates

Models are loaded, but problem with multiple materials and stuff

Assimp model loader, but multiple materials are not loaded

Solved the problem with shaders, normals, bumps and so on, awesome job. Now need just to rewrite the spot shader because it is wrong and ugly

Fixed bug with sponza, but problems with shader refactoring. Fixing tomorrow

Fixed problem with specular shader

Fixed last problem with shader

fix